requirements.txt
langchain
python-dotenv
langchain-community
langchain-openai
mysql-connector-python
streamlit
streamlit-chat